1. Replaces some uses of int with size_t where needed. 2. Added the WinSock.h include because windows is strange and puts it's timeval in there.

+++ bits/SpaceGroup/PCPort.cpp
@@ -1,14 +1,13 @@
-//#include "stdafx.h"
-
 #if defined(_WIN32)
 #include <string.h>
 #include "PCPort.h"
 #include <sys/types.h>
 #include <sys/timeb.h>
+//#include <WinSock.h>
 
-void bzero(void* pPoint, long pBytes)
+void bzero(void* pPoint, size_t pBytes)
 {
-	for (long i = 0; i < pBytes; i++)
+	for (size_t i = 0; i < pBytes; i++)
 	{
 		((char*)pPoint)[i] = 0;
 	}
